DopeEvents/events/models.py
from django.db import models
from django.contrib.auth.models import AbstractUser
from django.urls import reverse
from django.utils import timezone
from django.conf import settings

# Import merchandise models
from .models_merchandise import MerchandiseCategory, Merchandise, MerchandiseOrder, OrderItem

# Conditionally import Cloudinary
if hasattr(settings, 'CLOUDINARY_STORAGE') and settings.CLOUDINARY_STORAGE:
    import cloudinary.models
    ProfilePictureField = cloudinary.models.CloudinaryField('profile_pics', blank=True, null=True)
else:
    ProfilePictureField = models.ImageField(upload_to='profile_pics/', blank=True, null=True)

class User(AbstractUser):
    # User type fields
    is_buyer = models.BooleanField(default=False)
    is_seller = models.BooleanField(default=False)
    is_pro = models.BooleanField(default=False)
    phone_number = models.CharField(max_length=20, blank=True)
    profile_picture = ProfilePictureField
    
    # Seller specific fields
    business_name = models.CharField(max_length=100, blank=True)
    business_description = models.TextField(blank=True)
    
    groups = models.ManyToManyField(
        'auth.Group',
        related_name='custom_user_set',
        blank=True,
        verbose_name='groups',
        help_text='The groups this user belongs to.'
    )
    user_permissions = models.ManyToManyField(
        'auth.Permission',
        related_name='custom_user_set',
        blank=True,
        verbose_name='user permissions',
        help_text='Specific permissions for this user.'
    )

    @property
    def has_active_subscription(self):
        return hasattr(self, 'subscription') and self.subscription.is_active()

    def get_dashboard_url(self):
        if self.is_seller:
            return reverse('dashboard')
        elif self.is_buyer:
            return reverse('home')
        return reverse('home')

class Category(models.Model):
    name = models.CharField(max_length=100, null=False, unique=True)
    description = models.TextField(blank=True)
    slug = models.SlugField(max_length=100, unique=True)
    created_at = models.DateTimeField(auto_now_add=True)

    class Meta:
        verbose_name_plural = 'Categories'
        ordering = ['name']

    def __str__(self):
        return self.name

    def get_absolute_url(self):
        return reverse('category_detail', kwargs={'slug': self.slug})

class Event(models.Model):
    organizer = models.ForeignKey(User, on_delete=models.CASCADE, related_name='events')
    category = models.ForeignKey(Category, on_delete=models.CASCADE, null=True, related_name='events')
    title = models.CharField(max_length=150)
    description = models.TextField()
    image = models.ImageField(upload_to='event_images/', blank=True, null=True)
    date = models.DateTimeField()
    location = models.CharField(max_length=300)
    total_tickets = models.PositiveIntegerField(default=0)
    available_tickets = models.PositiveIntegerField(default=0)
    created_at = models.DateTimeField(auto_now=True)
    updated_at = models.DateTimeField(auto_now=True)
    is_active = models.BooleanField(default=True)

    class Meta:
        ordering = ['-created_at']

    def __str__(self):
        return self.title

    def get_absolute_url(self):
        return reverse('event_detail', kwargs={'pk': self.pk})

    @property
    def tickets_sold(self):
        """Return total number of tickets sold across all categories"""
        return sum(
            category.tickets_sold 
            for category in self.ticket_categories.all()
        )

    @property
    def is_sold_out(self):
        """Check if any tickets are available in any category"""
        # If no ticket categories exist, don't show as sold out
        if not self.ticket_categories.exists():
            return False
        
        # Check if any category has available tickets within sales period
        available_categories = self.get_available_categories()
        return not available_categories.exists()

    @property
    def is_past_event(self):
        return self.date < timezone.now()

    @property
    def lowest_ticket_price(self):
        category = self.ticket_categories.order_by('price').first()
        return category.price if category else None

    @property
    def highest_ticket_price(self):
        category = self.ticket_categories.order_by('-price').first()
        return category.price if category else None

    def get_available_categories(self):
        """Get all available ticket categories for this event"""
        now = timezone.now()
        return self.ticket_categories.filter(
            available_tickets__gt=0,
            sales_start__lte=now,
            sales_end__gte=now
        )

    def get_total_revenue(self):
        """Calculate total revenue from all ticket categories"""
        return sum(category.get_revenue() for category in self.ticket_categories.all())

class TicketCategory(models.Model):
    CATEGORY_TYPES = [
        ('regular', 'Regular'),
        ('vip', 'VIP'),
        ('group', 'Group'),
        ('early_bird', 'Early Bird'),
    ]

    event = models.ForeignKey(Event, on_delete=models.CASCADE, related_name='ticket_categories')
    name = models.CharField(max_length=100)
    category_type = models.CharField(max_length=20, choices=CATEGORY_TYPES)
    price = models.DecimalField(max_digits=10, decimal_places=2)
    initial_tickets = models.PositiveIntegerField(
        default=0,
        help_text="Initial number of tickets in this category"
    )
    available_tickets = models.PositiveIntegerField(
        default=0,
        help_text="Current number of tickets available"
    )
    description = models.TextField(blank=True)
    max_tickets_per_purchase = models.PositiveIntegerField(
        default=10,
        help_text="Maximum number of tickets one person can buy"
    )
    sales_start = models.DateTimeField()
    sales_end = models.DateTimeField()

    class Meta:
        unique_together = ['event', 'category_type']
        ordering = ['price']

    def __str__(self):
        return f"{self.event.title} - {self.name} (${self.price})"

    def save(self, *args, **kwargs):
        if not self.pk and self.available_tickets:
            self.initial_tickets = self.available_tickets
        super().save(*args, **kwargs)

    @property
    def tickets_sold(self):
        if self.initial_tickets is None or self.available_tickets is None:
            return 0
        return max(0, self.initial_tickets - self.available_tickets)

    def get_sales_percentage(self):
        if not self.initial_tickets:
            return 0
        return min(100, (self.tickets_sold * 100) // self.initial_tickets)

    @property
    def is_available(self):
        """Check if tickets are available and within sales period"""
        now = timezone.now()
        return (
            self.available_tickets > 0 and
            self.sales_start <= now <= self.sales_end
        )

    @property
    def total_tickets(self):
        return self.initial_tickets

    def get_revenue(self):
        return sum(ticket.total_amount for ticket in self.tickets.all())

class Ticket(models.Model):
    STATUS_CHOICES = [
        ('pending', 'Pending'),
        ('confirmed', 'Confirmed'),
        ('cancelled', 'Cancelled'),
        ('used', 'Used'),
    ]

    event = models.ForeignKey(Event, on_delete=models.CASCADE, related_name='tickets')
    ticket_category = models.ForeignKey(TicketCategory, on_delete=models.CASCADE, null=True, blank=True, related_name='tickets')
    buyer = models.ForeignKey(User, on_delete=models.CASCADE, related_name='purchased_tickets',null=True,blank=True)
    buyer_name = models.CharField(max_length=100)
    buyer_email = models.EmailField()
    buyer_phone = models.CharField(max_length=20, blank=True)
    quantity = models.PositiveIntegerField(default=1)
    unit_price = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
    total_amount = models.DecimalField(max_digits=10, decimal_places=2)
    stripe_payment_intent_id = models.CharField(max_length=200, blank=True)
    purchased_at = models.DateTimeField(auto_now_add=True)
    ticket_code = models.CharField(max_length=50, unique=True)
    transaction_code = models.CharField(max_length=50, blank=True, null=True)
    status = models.CharField(max_length=20, choices=STATUS_CHOICES, default='pending')
    used_at = models.DateTimeField(null=True, blank=True)
    cancelled_at = models.DateTimeField(null=True, blank=True)

    def save(self, *args, **kwargs):
        if not self.unit_price:
            self.unit_price = self.ticket_category.price
        self.total_amount = self.unit_price * self.quantity
        if not self.ticket_code:
            import uuid
            self.ticket_code = str(uuid.uuid4())[:8].upper()
        super().save(*args, **kwargs)

    def mark_as_used(self):
        self.status = 'used'
        self.used_at = timezone.now()
        self.save()

    def cancel(self):
        self.status = 'cancelled'
        self.cancelled_at = timezone.now()
        self.save()

class Subscription(models.Model):
    SUBSCRIPTION_PLANS = [
        ('basic', 'Basic'),
        ('pro', 'Professional'),
        ('premium', 'Premium'),
    ]

    user = models.OneToOneField(User, on_delete=models.CASCADE)
    stripe_customer_id = models.CharField(max_length=100)
    stripe_subscription_id = models.CharField(max_length=100)
    plan = models.CharField(max_length=20, choices=SUBSCRIPTION_PLANS)
    status = models.CharField(max_length=20)
    created_at = models.DateTimeField(auto_now_add=True)
    expires_at = models.DateTimeField()

    def is_active(self):
        return self.status == 'active' and self.expires_at > timezone.now()