DopeEvents/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.9.0
attrs==25.3.0
boto3==1.29.7
botocore==1.32.7
certifi==2025.6.15
charset-normalizer==3.4.2
Django==4.2.7
django-storages==1.14.2
frozenlist==1.7.0
idna==3.10
jmespath==1.0.1
multidict==6.6.3
pillow==11.3.0
propcache==0.3.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.4
s3transfer==0.8.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stripe==7.8.0
twilio==8.10.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.0.7
wheel==0.45.1
yarl==1.20.1